Start a CODEOWNERS file
I think this is a reasonably sane way to do access control. As a trial,
I'm just going to set it up to request reviews from the listed people,
but not to require them. We may want to make it a requirement in future.

Thus far I've added what I think is correct for `plutus`, `ledger`,
`ouroboros-consensus`, and the patched pacakges. I don't know who or what to put for
`cardano-node`, ideally there would be a team.

It would be great to find people for the other repositories in here.

I've also added some commentary to the README suggesting when people
should add themselves to CODEOWNERS.

Work towards #33.