Add a builder for packages using Nix, set cardano-node as a smoke test
This adds the infrastructure to build packages from CHaP with Nix in our
CI. There is a decent amount of ugliness but it works.

At the moment it doesn't try to be particularly clever: you can build a
single package-version, or a bunch of them (better for parallelism!), and
it's up to you to tell it what to build. But that's relatively easy to
do externally once the nix stuff is in place.