Improve scripts/add-revions.sh
The need to pass a `BUILT_REPO` arg is very confusing and it is small enough to just DL the latest to a temp dir.

This change also only uses the `BUILT_REPO` if there is no existing revision.  If there is a revision it copies the `.cabal` file from there (in case it is not in the BUILT_REPO yet).

Also uses `date -u` instead of `date --utc` so that we don't have to install `gdate` on macOS to make it work.