Add marlowe-runtime to smoke tests
Rationale: marlowe-runtime is an excellent smoke test. It is fairly far
downstream, and it depends on several other packages in the registry. As
such, it breaks easily if upstream packages fail to specify their
dependency version constraints accurately. Additionally, it is a
significant amount of work to fix the build each time new marlowe
packages are released, because updates to other packages frequently break
it in the meantime, so keeping it always building will make this much
easier.