Add marlowe-plutus-0.1.0.0 (#499)
* Add marlowe-runtime to smoke tests

Rationale: marlowe-runtime is an excellent smoke test. It is fairly far
downstream, and it depends on several other packages in the registry. As
such, it breaks easily if upstream packages fail to specify their
dependency version constraints accurately. Additionally, it is a
significant amount of work to fix the build each time new marlowe
packages are released, because updates to other packages frequently break
it in the meantime, so keeping it always building will make this much
easier.

* Added marlowe-plutus-0.1.0.0

From https://github.com/input-output-hk/marlowe-plutus at b7639be1239a589f8965828519c97542f60ecf99

* Update codeowners and exclude GHC9 for marlowe-plutus

* Remove marlowe-runtime smoke test package